--- a/HPXMLtoOpenStudio/measure.xml
+++ b/HPXMLtoOpenStudio/measure.xml
@@ -3,8 +3,8 @@
   <schema_version>3.0</schema_version>
   <name>hpxm_lto_openstudio</name>
   <uid>b1543b30-9465-45ff-ba04-1d1f85e763bc</uid>
-  <version_id>c26aaad6-ee0c-404d-8f7b-fee3cccf2f06</version_id>
-  <version_modified>20200513T150005Z</version_modified>
+  <version_id>73af152a-06f2-4120-ac67-bdbabd96dcbc</version_id>
+  <version_modified>20200513T154137Z</version_modified>
   <xml_checksum>D8922A73</xml_checksum>
   <class_name>HPXMLtoOpenStudio</class_name>
   <display_name>HPXML to OpenStudio Translator</display_name>
@@ -489,7 +489,7 @@
       <filename>measure.rb</filename>
       <filetype>rb</filetype>
       <usage_type>script</usage_type>
-      <checksum>FA27B272</checksum>
+      <checksum>A2072F8C</checksum>
     </file>
   </files>
 </measure>