--- a/HPXMLtoOpenStudio/resources/hpxml.rb
+++ b/HPXMLtoOpenStudio/resources/hpxml.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require_relative 'xmlhelper'
-
 '''
 Example Usage: