--- a/HPXMLtoOpenStudio/resources/location.rb
+++ b/HPXMLtoOpenStudio/resources/location.rb
@@ -1,9 +1,5 @@
 # frozen_string_literal: true
 
-require_relative 'constants'
-require_relative 'unit_conversions'
-require_relative 'weather'
-
 class Location
   def self.apply(model, runner, weather_file_path, weather_cache_path, dst_start_date, dst_end_date)
     weather, epw_file = apply_weather_file(model, runner, weather_file_path, weather_cache_path)