--- a/HPXMLtoOpenStudio/resources/xmlhelper.rb
+++ b/HPXMLtoOpenStudio/resources/xmlhelper.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require 'oga'
-
 class XMLHelper
   # Adds the child element with 'element_name' and sets its value. Returns the
   # child element.