Attempt to speed up the workflow by optimizing require calls.

Local testing w/ base.xml shows the following results:

  | master | this branch
-- | -- | --
Windows | 8.8s | 8.1s
Linux (WSL) | 6.9s | 6.3s

Note: The CI is largely unaffected by this change due to the way that we run tests within a single ruby process, resulting in gems being cached between simulations.